Create purchase history for shop items.
Once the user claims a reward, the user's points will be deducted by the price if they have enough. Then it will create a new item within purchase history. The calls for creating a new purchase history item and getting all history is working. Just need to add the route in App.jsx and a button for it.